shaders added

+++ inc/kernals/attention.h
@@ -2,7 +2,7 @@
 #define __ATTENTION_H__
 #include<stdint.h>
 #include<stdbool.h>
-void attention(const double* __restrict Q, const double* __restrict K, 
-    const double* __restrict V,const uint32_t n, const uint32_t d, double*__restrict out, 
+void attention(const float* __restrict Q, const float* __restrict K, 
+    const float* __restrict V,const uint32_t n, const uint32_t d, float*__restrict out, 
     const bool masked);
 #endif
