+++ inc/kernals/sigmoid.h
@@ -1,5 +1,5 @@
 #ifndef __SIGMOID_H__
 #define __SIGMOID_H__
 #include<stdint.h>
-void sigmoid(const double* __restrict x, double* __restrict y,const uint32_t len);
+void sigmoid(const float* __restrict x, float* __restrict y,const uint32_t len);
 #endif
