+++ inc/kernals/silu.h
@@ -1,5 +1,5 @@
 #ifndef __SILU_H__
 #define __SILU_H__
 #include<stdint.h>
-void SILU(const double* __restrict x, double* __restrict y, const uint32_t len);
+void SILU(const float* __restrict x, float* __restrict y, const uint32_t len);
 #endif
