+++ src/kernals/silu.c
@@ -1,8 +1,8 @@
 #include"../../inc/kernals/silu.h"
 #include <math.h>
 
-void SILU(const double* __restrict x, double* __restrict y, const uint32_t len){
+void SILU(const float* __restrict x, float* __restrict y, const uint32_t len){
     for (int i = 0;i < len;i++) {
-        y[i] = x[i] / (1 + exp(-1 * x[i]));
+        y[i] = x[i] / (1.0f + expf(-1 * x[i]));
     }
 }
